+++ packages/natives/native/index.js
@@ -25,7 +25,7 @@ export const PtySession = nativeBindings.PtySession;
 export const Shell = nativeBindings.Shell;
 
 // functions
-export const __piNativesV0_2_6 = nativeBindings.__piNativesV0_2_6;
+export const __piNativesV0_2_7 = nativeBindings.__piNativesV0_2_7;
 export const applyBashFixups = nativeBindings.applyBashFixups;
 export const astEdit = nativeBindings.astEdit;
 export const astGrep = nativeBindings.astGrep;